Added new animations for Colour and Level Up
When the player increases to a new colour or level, the player data is shown and the Colour and Level is animated, followed by the container itself.
Added some debug stuff to track any weird updates to the player data.
Updated the cheat (solveBarOne) to add the game to played (as it was confusing the life out of the Colour and Level function ;)