Added difficulty check for medium puzzles
Some medium puzzles will have an area with 0 or 1 numbers showing in it.
This is slightly more difficult than I would like, so added a check to balance the puzzle out a bit by removing numbers from overloaded areas, and showing a random cell position where the area is under-populated.